refactor: 이미지 모듈 경로 수정 - #259
