+++ image/src/main/java/com/back/image/ImageApplication.java
@@ -1,4 +1,4 @@
-package com.back;
+package com.back.image;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
