+++ image/src/main/java/com/back/image/config/S3Config.java
@@ -1,4 +1,4 @@
-package com.back.config;
+package com.back.image.config;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
