+++ image/src/main/java/com/back/image/utils/ImageUtility.java
@@ -1,4 +1,4 @@
-package com.back.utils;
+package com.back.image.utils;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
