+++ image/src/test/java/com/back/app/usecase/UploadImageUseCaseTest.java
@@ -3,11 +3,11 @@ package com.back.app.usecase;
 import com.amazonaws.AmazonClientException;
 import com.amazonaws.AmazonServiceException;
 import com.amazonaws.services.s3.AmazonS3Client;
-import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.back.common.code.FailureCode;
 import com.back.common.exception.CustomException;
-import com.back.utils.ImageUtility;
+import com.back.image.app.usecase.UploadImageUseCase;
+import com.back.image.utils.ImageUtility;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
