+++ product/src/main/java/com/back/product/adapter/in/ApiV1ImageController.java
@@ -1,10 +1,10 @@
 package com.back.product.adapter.in;
 
-import com.back.app.ImageFacade;
-import com.back.dto.enums.ImageCategory;
+import com.back.image.app.ImageFacade;
+import com.back.image.dto.enums.ImageCategory;
 import com.back.common.code.SuccessCode;
 import com.back.common.response.CommonResponse;
-import com.back.dto.response.ImageUploadResponseDto;
+import com.back.image.dto.response.ImageUploadResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
